code_rag/requirements.txt
streamlit==1.50.0
pydantic-ai-slim[groq]==1.0.15
numpy==2.3.3
faiss-cpu==1.12.0
python-dotenv==1.1.1
scikit-learn==1.7.2
